[update] fp16 inference

+++ scripts/eval_toolcall.py
@@ -64,7 +64,7 @@ def init_model(args):
     else:
         model = AutoModelForCausalLM.from_pretrained(args.load_from, trust_remote_code=True)
     get_model_params(model, model.config)
-    return model.eval().to(args.device), tokenizer
+    return model.half().eval().to(args.device), tokenizer
 
 
 def parse_tool_calls(text):
